saque el protocolo del main
