+++ Prototipo/COMECA/COMECA/Debug/Core/Src/main.su
@@ -1,14 +1,9 @@
-main.c:138:5:main	8	static
-main.c:184:6:SystemClock_Config	88	static
-main.c:229:13:MX_TIM14_Init	8	static
-main.c:260:13:MX_GPIO_Init	40	static
-main.c:301:6:HAL_TIM_PeriodElapsedCallback	16	static
-main.c:308:6:HAL_GPIO_EXTI_Callback	16	static
-main.c:318:6:getCaliperData	16	static
-main.c:339:6:processBit	32	static
-main.c:350:6:onRisingEdgeOfReqSignal	16	static
-main.c:356:6:onRisingEdgeOfClockSignal	24	static
-main.c:378:20:digimaticGetMeasure	16	static
-main.c:388:6:analogValidate	16	static
-main.c:392:6:AnalogInDigitalOutManager	16	static
-main.c:422:6:Error_Handler	4	static,ignoring_inline_asm
+main.c:82:5:main	8	static
+main.c:128:6:SystemClock_Config	88	static
+main.c:173:13:MX_TIM14_Init	8	static
+main.c:204:13:MX_GPIO_Init	40	static
+main.c:245:6:HAL_TIM_PeriodElapsedCallback	16	static
+main.c:252:6:HAL_GPIO_EXTI_Callback	16	static
+main.c:263:6:analogValidate	16	static
+main.c:267:6:AnalogInDigitalOutManager	16	static
+main.c:297:6:Error_Handler	4	static,ignoring_inline_asm
