+++ Prototipo/COMECA/COMECA/Debug/makefile
@@ -61,8 +61,8 @@ all: main-build
 main-build: COMECA.elf secondary-outputs
 
 # Tool invocations
-COMECA.elf: $(OBJS) $(USER_OBJS) C:\DisenioBJAM\Prototipo\COMECA\COMECA\STM32F407VGTX_FLASH.ld makefile objects.list $(OPTIONAL_TOOL_DEPS)
-	arm-none-eabi-gcc -o "COMECA.elf" @"objects.list" $(USER_OBJS) $(LIBS) -mcpu=cortex-m4 -T"C:\DisenioBJAM\Prototipo\COMECA\COMECA\STM32F407VGTX_FLASH.ld" --specs=nosys.specs -Wl,-Map="COMECA.map" -Wl,--gc-sections -static --specs=nano.specs -mfpu=fpv4-sp-d16 -mfloat-abi=hard -mthumb -Wl,--start-group -lc -lm -Wl,--end-group
+COMECA.elf: $(OBJS) $(USER_OBJS) C:\Github_Repos\DisenioBJAM\Prototipo\COMECA\COMECA\STM32F407VGTX_FLASH.ld makefile objects.list $(OPTIONAL_TOOL_DEPS)
+	arm-none-eabi-gcc -o "COMECA.elf" @"objects.list" $(USER_OBJS) $(LIBS) -mcpu=cortex-m4 -T"C:\Github_Repos\DisenioBJAM\Prototipo\COMECA\COMECA\STM32F407VGTX_FLASH.ld" --specs=nosys.specs -Wl,-Map="COMECA.map" -Wl,--gc-sections -static --specs=nano.specs -mfpu=fpv4-sp-d16 -mfloat-abi=hard -mthumb -Wl,--start-group -lc -lm -Wl,--end-group
 	@echo 'Finished building target: $@'
 	@echo ' '
