+++ Prototipo/COMECA/COMECA/Debug/objects.list
@@ -1,3 +1,4 @@
+"Core/Src/digimatic.o"
 "Core/Src/main.o"
 "Core/Src/stm32f4xx_hal_msp.o"
 "Core/Src/stm32f4xx_it.o"
